% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBrazilCovid19Data.R
\name{getBrazilCovid19Data}
\alias{getBrazilCovid19Data}
\title{Get Covid-19 Brasil Data from Brasil.io}
\usage{
getBrazilCovid19Data(url)
}
\arguments{
\item{url}{String with the url}
}
\description{
This function retrieves the data from the https://www.brasil.io/home page.
}
