% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLastDataUpdateTimeBrasilIO.R
\name{getLastDataUpdateTimeBrasilIO}
\alias{getLastDataUpdateTimeBrasilIO}
\title{Return Last Time Brazil Covid-19 Cases Was Updated}
\usage{

  getLastDataUpdateTimeBrasilIO(url = "https://brasil.io/dataset/covid19/caso",
  xpath = "/html/body/div[1]/div/div[2]/div[2]/div[2]/p")
}
\arguments{
\item{url}{"https://brasil.io/dataset/covid19/caso"}

\item{xpath}{"/html/body/div[1]/div/div[2]/div[2]/div[2]/p"}
}
\value{
String with the data information
}
\description{
Return Last Time Brazil Covid-19 Cases Was Updated
}
