% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_mavg_states.R
\name{heatmap_mavg_states}
\alias{heatmap_mavg_states}
\title{Plot Heatmap of Moving Average Cases}
\usage{
heatmap_mavg_states(df)
}
\arguments{
\item{df}{Data Frame. Output from function getBrazilCovid19Data()}
}
\value{
echarts4r plot
}
\description{
Plot Heatmap of Moving Average Cases
}
