% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cities_in_state_cases_over_time.R
\name{plot_cities_in_state_cases_over_time}
\alias{plot_cities_in_state_cases_over_time}
\title{Plot the Evolution Within Cities of a Selected State}
\usage{
plot_cities_in_state_cases_over_time(df_covid19, State)
}
\arguments{
\item{df_covid19}{Data Frame Returned from getBrazilCovid19Data()}

\item{State}{String with selected state (input$selected_state)}
}
\description{
Plot the Evolution Within Cities of a Selected State
}
