% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confirmed_per_million.R
\name{plot_confirmed_per_million}
\alias{plot_confirmed_per_million}
\title{Title}
\usage{
plot_confirmed_per_million(df, ln_confirmed = F)
}
\arguments{
\item{df}{}

\item{ln_confirmed}{}
}
\description{
Title
}
