% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_covid19_br_map.R
\name{plot_covid19_br_map}
\alias{plot_covid19_br_map}
\title{Plots Covid-19 Per State in Brazil Map}
\usage{
plot_covid19_br_map(covid19_df, polygon_df)
}
\arguments{
\item{covid19_df}{Data Frame resulted from function get_covid19_br_data()}

\item{polygon_df}{Shape files resultado geobr::state(year==2018)}
}
\value{
leaflet plot
}
\description{
Plots Covid-19 Per State in Brazil Map
}
