% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_death_confirmed_over_time_states.R
\name{plot_death_confirmed_over_time_states}
\alias{plot_death_confirmed_over_time_states}
\title{Plot Death Rate per State Over Time}
\usage{
plot_death_confirmed_over_time_states(df_covid19)
}
\arguments{
\item{df_covid19}{Data Frame Returned from getBrazilCovid19Data()}
}
\description{
Plot Death Rate per State Over Time
}
