% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_moving_avg.R
\name{plot_moving_avg}
\alias{plot_moving_avg}
\title{Plot New Cases and New Deaths Moving Average}
\usage{
plot_moving_avg(df, state_view = F)
}
\arguments{
\item{df}{Data Frame returned from getBrazilCovid19Data() function.}

\item{state_view}{Logical. If F (deafault) then return the view for the whole country. 
If T, the return the state view}
}
\value{
ggplot object
}
\description{
This function returns the plot of the 7 day moving average of new cases and new deaths.
}
