% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_states_mavg_new_cases.R
\name{plot_multiple_states_mavg_new_cases}
\alias{plot_multiple_states_mavg_new_cases}
\title{Plot Multiples States new cases over time}
\usage{
plot_multiple_states_mavg_new_cases(df, i_ini, i_final)
}
\arguments{
\item{df}{Data Frame. Output from getBrazilCovid19Data() function.}

\item{i_ini}{Integer. index of unique(df$state) to start.}

\item{i_final}{Integer. Index of unique(df$state) to end}
}
\description{
The objective of this function is to make it possible to 
check the new cases trend of the states
}
