% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cases_state_map.R
\name{plot_cases_state_map}
\alias{plot_cases_state_map}
\title{Plots the Cases for Each City Given a Selected State}
\usage{
plot_cases_state_map(df_covid19, State, state_shape_files)
}
\arguments{
\item{df_covid19}{Data Frame Returned from getBrazilCovid19Data()}

\item{State}{String with selected state (input$selected_state)}

\item{state_shape_files}{Shape files of the states}
}
\description{
Plots the Cases for Each City Given a Selected State
}
